+++ catapult/api.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (C) 2022 Osmo Salomaa
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+"""API for plugins."""
+
+from catapult.plugin import Plugin # noqa
+from catapult.preferences import PreferencesItem # noqa
+from catapult.search import SearchResult # noqa
+from catapult.util import copy_text_to_clipboard # noqa
+from catapult.util import get_desktop_environment # noqa
+from catapult.util import is_path # noqa
+from catapult.util import is_uri # noqa
+from catapult.util import lookup_icon # noqa